Delete the rest of the progress v1 page
