--- apps/src/templates/sectionProgress/SectionProgress.jsx
@@ -1,289 +1,0 @@
-import PropTypes from 'prop-types';
-import React, {Component} from 'react';
-import {connect} from 'react-redux';
-
-import FontAwesome from '@cdo/apps/legacySharedComponents/FontAwesome';
-import logToCloud from '@cdo/apps/logToCloud';
-import {EVENTS, PLATFORMS} from '@cdo/apps/metrics/AnalyticsConstants';
-import analyticsReporter from '@cdo/apps/metrics/AnalyticsReporter';
-import {
-  getSelectedCourseId,
-  getSelectedUnitPosition,
-  setUnit,
-} from '@cdo/apps/redux/unitSelectionRedux';
-import ProgressTableView from '@cdo/apps/templates/sectionProgress/progressTables/ProgressTableView';
-import SortByNameDropdown from '@cdo/apps/templates/SortByNameDropdown';
-import i18n from '@cdo/locale';
-
-import {h3Style} from '../../legacySharedComponents/Headings';
-
-import {ViewType, unitDataPropType} from './sectionProgressConstants';
-import {loadUnitProgress} from './sectionProgressLoader';
-import {getCurrentUnitData, setLessonOfInterest} from './sectionProgressRedux';
-import UnitSelector from './UnitSelector';
-
-import styleConstants from './progressTables/progress-table-constants.module.scss';
-import navigationStyles from '@cdo/apps/templates/teacherNavigation/teacher-navigation.module.scss';
-
-const SECTION_PROGRESS = 'SectionProgress';
-
-/**
- * Given a particular section, this component owns figuring out which script to
- * show progress for (selected via a dropdown), and querying the server for
- * student progress. Child components then have the responsibility for displaying
- * that progress.
- */
-class SectionProgress extends Component {
-  static propTypes = {
-    //Provided by redux
-    scriptId: PropTypes.number,
-    courseVersionId: PropTypes.number,
-    courseId: PropTypes.number,
-    unitPosition: PropTypes.number,
-    sectionId: PropTypes.number,
-    currentView: PropTypes.oneOf(Object.values(ViewType)),
-    scriptData: unitDataPropType,
-    setUnit: PropTypes.func.isRequired,
-    setLessonOfInterest: PropTypes.func.isRequired,
-    isLoadingProgress: PropTypes.bool.isRequired,
-    isRefreshingProgress: PropTypes.bool,
-  };
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      reportedInitialRender: false,
-      loadedSectionId: null,
-    };
-  }
-
-  componentDidMount() {
-    if (this.props.scriptId && this.props.courseId && this.props.unitPosition) {
-      loadUnitProgress(
-        this.props.scriptId,
-        this.props.sectionId,
-        this.props.courseId,
-        this.props.unitPosition
-      )?.then(() => {
-        this.setState(state => ({
-          ...state,
-          loadedSectionId: this.props.sectionId,
-        }));
-      });
-    }
-  }
-
-  componentDidUpdate(prevProps) {
-    if (
-      prevProps.scriptId !== this.props.scriptId ||
-      prevProps.sectionId !== this.props.sectionId ||
-      prevProps.courseId !== this.props.courseId ||
-      prevProps.unitPosition !== this.props.unitPosition
-    ) {
-      loadUnitProgress(
-        this.props.scriptId,
-        this.props.sectionId,
-        this.props.courseId,
-        this.props.unitPosition
-      )?.then(() => {
-        this.setState(state => ({
-          ...state,
-          loadedSectionId: this.props.sectionId,
-        }));
-      });
-    }
-
-    if (this.levelDataInitialized() && !this.state.reportedInitialRender) {
-      logToCloud.addPageAction(
-        logToCloud.PageAction.SectionProgressRenderedWithData,
-        {
-          sectionId: this.props.sectionId,
-          scriptId: this.props.scriptId,
-        }
-      );
-      this.setState(state => ({
-        ...state,
-        reportedInitialRender: true,
-      }));
-    }
-
-    if (
-      (prevProps.scriptId !== this.props.scriptId ||
-        prevProps.sectionId !== this.props.sectionId ||
-        prevProps.isLoadingProgress !== this.props.isLoadingProgress ||
-        prevProps.isRefreshingProgress !== this.props.isRefreshingProgress) &&
-      !this.props.isLoadingProgress &&
-      !this.props.isRefreshingProgress
-    ) {
-      analyticsReporter.sendEvent(
-        EVENTS.PROGRESS_VIEWED_FIXED,
-        {
-          sectionId: this.props.sectionId,
-          unitId: this.props.scriptId,
-          windowWidth: window.innerWidth,
-          windowHeight: window.innerHeight,
-        },
-        PLATFORMS.STATSIG
-      );
-    }
-  }
-
-  onChangeScript = (scriptId, courseVersionId) => {
-    this.props.setUnit(scriptId, courseVersionId);
-
-    analyticsReporter.sendEvent(
-      EVENTS.PROGRESS_CHANGE_UNIT,
-      {
-        sectionId: this.props.sectionId,
-        oldUnitId: this.props.scriptId,
-        unitId: scriptId,
-      },
-      PLATFORMS.STATSIG
-    );
-  };
-
-  onChangeLevel = lessonOfInterest => {
-    this.props.setLessonOfInterest(lessonOfInterest);
-
-    analyticsReporter.sendEvent(EVENTS.PROGRESS_JUMP_TO_LESSON, {
-      sectionId: this.props.sectionId,
-      unitId: this.props.scriptId,
-      lesson: this.props.scriptData.lessons[lessonOfInterest].id,
-    });
-  };
-
-  levelDataInitialized = () => {
-    const {
-      scriptData,
-      isLoadingProgress,
-      isRefreshingProgress,
-      sectionId,
-      scriptId,
-    } = this.props;
-    const {loadedSectionId} = this.state;
-    return (
-      scriptData &&
-      !isLoadingProgress &&
-      !isRefreshingProgress &&
-      sectionId === loadedSectionId &&
-      scriptData.id === scriptId
-    );
-  };
-
-  render() {
-    const {currentView, scriptId, courseVersionId, scriptData, sectionId} =
-      this.props;
-    const levelDataInitialized = this.levelDataInitialized();
-    const showProgressTable =
-      levelDataInitialized &&
-      (currentView === ViewType.SUMMARY || currentView === ViewType.DETAIL);
-
-    return (
-      <div
-        className={navigationStyles.widthLockedPage}
-        // eslint-disable-next-line react/forbid-dom-props
-        data-testid="section-progress-v1"
-      >
-        <div style={styles.topRowContainer}>
-          <div>
-            <div style={{...h3Style, ...styles.heading}}>
-              {i18n.selectACourse()}
-            </div>
-            <UnitSelector
-              scriptId={scriptId}
-              courseVersionId={courseVersionId}
-              onChange={this.onChangeScript}
-            />
-          </div>
-        </div>
-        <div style={styles.topRowContainer}>
-          {showProgressTable && (
-            <div style={styles.sortOrderSelect}>
-              <SortByNameDropdown
-                sectionId={sectionId}
-                unitName={scriptData?.title}
-                source={SECTION_PROGRESS}
-              />
-            </div>
-          )}
-        </div>
-
-        <div style={{clear: 'both'}}>
-          {!levelDataInitialized && (
-            <FontAwesome
-              id="uitest-spinner"
-              icon="spinner"
-              className="fa-pulse fa-3x"
-            />
-          )}
-          {showProgressTable && <ProgressTableView currentView={currentView} />}
-        </div>
-      </div>
-    );
-  }
-}
-
-const sortOrderMargin = 15;
-
-const styles = {
-  heading: {
-    marginBottom: 0,
-  },
-  topRowContainer: {
-    display: 'flex',
-    alignItems: 'flex-end',
-    marginBottom: 10,
-    width: '100%',
-  },
-  chevronLink: {
-    display: 'flex',
-    flex: 1,
-    justifyContent: 'flex-end',
-  },
-  icon: {
-    paddingRight: 5,
-  },
-  toggle: {
-    margin: '0px 30px',
-  },
-  show: {
-    display: 'block',
-  },
-  hide: {
-    display: 'none',
-  },
-  studentTooltip: {
-    display: 'flex',
-    textAlign: 'center',
-  },
-  sortOrderSelect: {
-    marginRight: sortOrderMargin,
-    marginBottom: 8,
-    width: parseInt(styleConstants.STUDENT_LIST_WIDTH) - sortOrderMargin,
-  },
-};
-
-export const UnconnectedSectionProgress = SectionProgress;
-
-export default connect(
-  state => ({
-    scriptId: state.unitSelection.scriptId,
-    courseVersionId: state.unitSelection.courseVersionId,
-    courseId: getSelectedCourseId(state),
-    unitPosition: getSelectedUnitPosition(state),
-    sectionId: state.teacherSections.selectedSectionId,
-    currentView: state.sectionProgress.currentView,
-    scriptData: getCurrentUnitData(state),
-    isLoadingProgress: state.sectionProgress.isLoadingProgress,
-    isRefreshingProgress: state.sectionProgress.isRefreshingProgress,
-  }),
-  dispatch => ({
-    setUnit(scriptId, courseVersionId) {
-      dispatch(setUnit(scriptId, courseVersionId));
-    },
-    setLessonOfInterest(lessonOfInterest) {
-      dispatch(setLessonOfInterest(lessonOfInterest));
-    },
-  })
-)(SectionProgress);
