--- apps/src/templates/sectionProgress/progressTables/progress-table-constants.module.scss
@@ -1,21 +1,0 @@
-@use "sass:math";
-@import "style-constants";
-
-$scrollbar-height: 15px;
-$row-height: 42px;
-$expanded-row-height: 30px;
-$max-rows: 14;
-$max-height: $max-rows * $row-height + $scrollbar-height;
-$table-width: calc($content-width - 64px);
-$student-list-width: 170px;
-$content-view-width: $table-width - $student-list-width;
-
-:export {
-  MAX_BODY_HEIGHT: math.div($max-height, 1px);
-  TABLE_WIDTH: math.div($table-width, 1px);
-  STUDENT_LIST_WIDTH: math.div($student-list-width, 1px);
-  CONTENT_VIEW_WIDTH: math.div($content-view-width, 1px);
-  ROW_HEIGHT: math.div($row-height, 1px);
-  MAX_ROWS: $max-rows;
-  MIN_COLUMN_WIDTH: 40;
-}
