--- apps/src/templates/sectionProgress/progressTables/progressTableStyles.scss
@@ -1,131 +1,0 @@
-@use "./progress-table-constants.module.scss" as constants;
-@import "color.scss";
-@import "style-constants";
-
-.progress-table {
-  table {
-    table-layout: fixed;
-    border: 1px solid;
-    border-color: $border-gray;
-    border-collapse: separate;
-  }
-  th {
-    background-color: $table-header;
-    color: $charcoal;
-    border-width: 0 1px 2px 0;
-    border-color: $border-gray;
-    height: constants.$row-height;
-  }
-  th,
-  td {
-    padding: 0;
-    box-sizing: border-box;
-    border-right: 1px solid;
-    border-color: $border-gray;
-    &:last-child {
-      border-right: 0;
-    }
-  }
-  .cell-content {
-    padding: 0 4px;
-  }
-  .primary-row {
-    td {
-      height: constants.$row-height;
-    }
-  }
-  .expanded-row {
-    td {
-      height: constants.$expanded-row-height;
-      border-width: 0 1px 1px 0;
-      border-style: solid;
-      border-color: $border-gray;
-    }
-    td:last-of-type {
-      border-width: 0 0 1px 0;
-    }
-  }
-  .first-expanded-row {
-    td {
-      border-width: 1px 1px 1px 0;
-    }
-    td:last-of-type {
-      border-width: 1px 0;
-    }
-  }
-  .student-list {
-    table,
-    th,
-    td {
-      width: constants.$student-list-width;
-    }
-    tbody {
-      tr,
-      td {
-        display: block;
-        width: 100%;
-      }
-    }
-    .content {
-      height: constants.$row-height;
-      box-sizing: border-box;
-      padding: 10px;
-      font-size: 14px;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      overflow: hidden;
-      .fa {
-        html[dir="RTL"] & {
-          transform: scaleX(-1);
-        }
-      }
-    }
-    .expanded-row {
-      td {
-        border-width: 0 0 1px 0;
-      }
-    }
-    .first-expanded-row {
-      td {
-        border-width: 1px 0;
-      }
-    }
-  }
-  .dark-row {
-    background-color: $background_gray;
-  }
-  .content-view {
-    thead,
-    tbody {
-      max-width: constants.$content-view-width;
-      td {
-        text-align: center;
-      }
-    }
-    .lesson-arrow {
-      width: 100%;
-      display: flex;
-      align-items: center;
-      align-content: center;
-      .line {
-        width: 100%;
-        height: 0;
-        border: 1px solid;
-        margin-inline-end: -7px;
-        margin-inline-start: 4px;
-      }
-      .arrow {
-        display: inline-block;
-        border-style: solid;
-        border-width: 0 2px 2px 0;
-        width: 6px;
-        height: 6px;
-        transform: rotate(-45deg);
-        -webkit-transform: rotate(-45deg);
-        html[dir="RTL"] & {
-          border-width: 2px 0 0 2px;
-        }
-      }
-    }
-  }
-}
