--- apps/src/templates/sectionProgress/sectionProgressConstants.js
@@ -1,54 +1,0 @@
-import PropTypes from 'prop-types';
-
-import {studentType} from '@cdo/apps/templates/progress/progressTypes';
-
-/**
- * Shape for unitData
- * The data we get from the server's call to unit.summarize. The format
- * ends up being similar to that which we send to initProgress in progressRedux.
- * The important part is unitData.lessons, which gets used by levelsWithLesson
- * Note: unit was previously named script
- */
-export const unitDataPropType = PropTypes.shape({
-  id: PropTypes.number.isRequired,
-  csf: PropTypes.bool,
-  title: PropTypes.string,
-  path: PropTypes.string,
-  lessons: PropTypes.arrayOf(
-    PropTypes.shape({
-      levels: PropTypes.arrayOf(PropTypes.object).isRequired,
-    })
-  ),
-  family_name: PropTypes.string,
-  version_year: PropTypes.string,
-  name: PropTypes.string,
-});
-
-// Types of views of the progress tab
-export const ViewType = {
-  SUMMARY: 'summary', // lessons
-  DETAIL: 'detail', // levels
-};
-
-/**
- * @typedef {Object} StudentTableRow
- * An object wrapping around `studentType` used to manage the
- * expanded/collapsed state of section progress table rows.
- *
- * @property {string} id
- * @property {studentType} student
- * @property {number} expansionIndex 0 for the primary row,
- * otherwise 1-based index of expanded row
- * @property {bool} isExpanded used by primary row to track state
- * @property {bool} useDarkBackground used by primary row to determine background color
- */
-export const studentTableRowType = PropTypes.shape({
-  id: PropTypes.string.isRequired,
-  student: studentType.isRequired,
-  expansionIndex: PropTypes.number.isRequired,
-  isExpanded: PropTypes.bool,
-  useDarkBackground: PropTypes.bool,
-});
-
-export const tooltipIdForStudent = studentId =>
-  `tooltipIdForStudent${studentId}`;
