+++ apps/src/templates/sectionProgressV2/ExpandedProgressColumnHeader.jsx
@@ -5,11 +5,11 @@ import {connect} from 'react-redux';
 import i18n from '@cdo/locale';
 
 import FontAwesome from '../../legacySharedComponents/FontAwesome';
-import {removeExpandedLesson} from '../sectionProgress/sectionProgressRedux';
 
 import LessonTitleTooltip, {getTooltipId} from './LessonTitleTooltip';
 import {getLessonColumnHeaderId} from './LevelDataCell';
 import LevelProgressHeader from './LevelProgressHeader';
+import {removeExpandedLesson} from './sectionProgressRedux';
 
 import styles from './progress-table-v2.module.scss';
