+++ apps/src/templates/sectionProgressV2/sectionProgressLoader.js
