--- apps/test/unit/templates/sectionProgress/progressTables/ProgressTableStudentListTest.jsx
@@ -1,73 +1,0 @@
-import {shallow, mount} from 'enzyme'; // eslint-disable-line no-restricted-imports
-import React from 'react';
-import * as Sticky from 'reactabular-sticky';
-import * as Virtualized from 'reactabular-virtualized';
-
-import ProgressTableStudentList from '@cdo/apps/templates/sectionProgress/progressTables/ProgressTableStudentList';
-import {
-  fakeRowsForStudents,
-  fakeDetailRowsForStudent,
-} from '@cdo/apps/templates/sectionProgress/sectionProgressTestHelpers';
-import i18n from '@cdo/locale';
-
-const TEST_STUDENT_1 = {
-  id: 1,
-  name: 'Joe',
-};
-
-const TEST_STUDENT_2 = {
-  id: 2,
-  name: 'Jamie',
-};
-
-const STUDENT_ROWS = fakeRowsForStudents([TEST_STUDENT_1, TEST_STUDENT_2]);
-const DETAIL_ROWS = fakeDetailRowsForStudent(TEST_STUDENT_1);
-
-const DEFAULT_PROPS = {
-  sectionId: 1,
-  scriptData: {
-    id: 144,
-    name: 'csd1',
-  },
-  rows: STUDENT_ROWS,
-  onRow: () => {},
-  headers: ['Lesson'],
-  studentTimestamps: {3: 1610435096000, 4: 0},
-  onToggleRow: () => {},
-};
-
-const setUp = (overrideProps = {}) => {
-  const props = {...DEFAULT_PROPS, ...overrideProps};
-  return shallow(<ProgressTableStudentList {...props} />);
-};
-
-describe('ProgressTableStudentList', () => {
-  it('displays a header for each header in props.headers', () => {
-    const headers = ['Lesson', 'Level Type'];
-    const wrapper = setUp({headers});
-    const stickyHeaderComponent = wrapper.find(Sticky.Header);
-    expect(stickyHeaderComponent.contains(headers[0]));
-    expect(stickyHeaderComponent.contains(headers[1]));
-  });
-
-  it('displays a name for each student', () => {
-    const wrapper = setUp();
-    const studentRows = wrapper.find(Virtualized.Body).props().rows;
-    expect(studentRows).toHaveLength(2);
-    expect(studentRows.includes(STUDENT_ROWS[0])).toBe(true);
-    expect(studentRows.includes(STUDENT_ROWS[1])).toBe(true);
-  });
-
-  it('displays detail labels if detail rows are passed in', () => {
-    // reactabular initially only renders three rows, so we use a single
-    // student to avoid needing to workaround that.
-    const rows = [STUDENT_ROWS[0], ...DETAIL_ROWS];
-
-    // we need to use mount for the rows to actually be rendered
-    const wrapper = mount(
-      <ProgressTableStudentList {...DEFAULT_PROPS} rows={rows} />
-    );
-    expect(wrapper.contains(i18n.timeSpentMins())).toBe(true);
-    expect(wrapper.contains(i18n.lastUpdatedTitle())).toBe(true);
-  });
-});
