+++ apps/test/unit/templates/sectionProgressV2/ProgressTableV2Test.jsx
@@ -7,19 +7,17 @@ import unitSelection from '@cdo/apps/redux/unitSelectionRedux';
 import currentUser, {
   setSortByFamilyName,
 } from '@cdo/apps/templates/currentUserRedux';
+import ProgressTableV2 from '@cdo/apps/templates/sectionProgressV2/ProgressTableV2.jsx';
 import sectionProgress, {
   addExpandedLesson,
-} from '@cdo/apps/templates/sectionProgress/sectionProgressRedux';
-import {
-  createStore,
-  getScriptData,
-} from '@cdo/apps/templates/sectionProgress/sectionProgressTestHelpers';
-import ProgressTableV2 from '@cdo/apps/templates/sectionProgressV2/ProgressTableV2.jsx';
+} from '@cdo/apps/templates/sectionProgressV2/sectionProgressRedux';
 import teacherSections, {
   setStudentsForCurrentSection,
 } from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
 
 import {expect} from '../../../util/reconfiguredChai'; // eslint-disable-line no-restricted-imports
+
+import {createStore, getScriptData} from './sectionProgressTestHelpers';
 
 const STUDENT_1 = {id: 1, name: 'Student 1', familyName: 'FamNameB'};
 const STUDENT_2 = {id: 2, name: 'Student 2', familyName: 'FamNameA'};
